Use consistent email subject line for pipelines
When using the `gitlab_rails['gitlab_email_subject_suffix']`
configuration option it was not applied to all emails. This
commits add the support for email_subject_suffix to
pipeline notification emails. As a consequence of this change
the project names placement is consistent with other email subjects.
The project name is now at the beginning of the subject for pipeline
emails aswell.

Changelog: changed

Issue: https://gitlab.com/gitlab-org/gitlab/-/issues/213383