Fix Security Configuration enable via MR buttons
This makes sure that the "Configure via Merge Request" button is
displayed for Secret Detection in the [new][1] Security Configuration
page for all tiers.

The button was not working in FOSS since the client-side mutation was
only defined in EE, so that was moved to FOSS. The backend
implementation of the mutation is already in FOSS.

This also ensures that a similar button is displayed for Dependency
Scanning in Ultimate projects when the sec_dependency_scanning_ui_enable
feature flag is enabled.

Addresses https://gitlab.com/gitlab-org/gitlab/-/issues/326005 and
https://gitlab.com/gitlab-org/gitlab/-/issues/328426.

[1]: https://gitlab.com/gitlab-org/gitlab/-/issues/326926