Include position for cache key of DiffDiscussion
In https://gitlab.com/gitlab-org/gitlab/-/issues/332967, we
introduced caching on discussions endpoint.

Upon testing it further while rolling it out, found a bug wherein
the diff note that should still be displayed won't display anymore
because position changed. The position stored in cache no longer
applies as it's outdated.

With this fix, we include the position of the first note in diff
discussion (as sha) in the cache_key. This way, when the position
changes, the cache key will be updated and will result to a cache
miss and will return updated data.