Drop support for Safari 13.0
When we [recently updated][0] our supported browser versions, we dropped
support for Safari < 13.0. Our docs state that we could drop support for
Safari 13.0 as well, while still supporting 13.1.

Safari 13.0 is barely used (and End of Life anyway) and dropping support
for it will allow us to remove polyfills for the [ResizeObserver API][1]
andto use the newer [Clipboard APIs][2].

[0]: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/63994
[1]: https://developer.mozilla.org/en-US/docs/Web/API/ResizeObserver
[2]: https://developer.mozilla.org/en-US/docs/Web/API/Clipboard

Changelog: changed