Look up SSH keys by SHA256 fingerprint, not MD5 fingerprint
This change takes GitLab one step closer to being FIPS-compliant, and
means git operations over SSH can work in a FIPS-enforcing environment.

GitLab Shell sends the full SSH key to GitLab, and we have the SHA256
fingerprint stored in the database already, so which fingerprint we use
to look up the key is an implementation detail; the tests (correctly,
IMO) don't mandate a particular form.

The next step for FIPS compliance on this feature is to stop generating
MD5 fingerprints when SSH keys are added.

Changelog: fixed