Remove Database::Connection.healthy?
This method was added using an EE only method, and just inverted the
result of Postgresql::ReplicationSlot. This however creates a problem: a
Connection is tied to a specific database connection, but
ReplicationSlot is tied to the main database.

When we add support for multiple databases to ReplicationSlot, we'd
somehow have to figure out what model to use for the `healthy?` class,
as there would a ReplicationSlot model per primary database. Instead of
trying to work this out, which likely would get hairy, we simply remove
the `healthy?` method in favour of using ReplicationSlot directly. It
was only used in one place anyway, ant there is no harm in using
ReplicationSlot directly in that instance.