Use http read total timeout by default
We want to work towards using read total timeout by default.
To make sure it does not unintentionally break something, this
keeps the use_read_total_timeout option and only raises an
error if it is true. Otherwise, it logs the error. After monitoring
for unexpected errors, we can remove the option and make raising
an error, the default behavior.