Allow missing side when listing conflicts on diffs
In https://gitlab.com/gitlab-org/gitaly/-/merge_requests/3648,
support for listing conflicts even when there's a missing side has
been added to ListConflictFiles RPC. This only happens when
`allow_tree_conflicts` is set to `true` in the RPC request.

This allows gitlab-rails to use that parameter when displaying
MR diffs with conflicts. The `allow_tree_conflicts` will be set
to `true` only when `display_merge_conflicts_in_diff` is enabled.
Otherwise, it'll be set to `false`.