workers: Remove support for asynchronous remote removal
There are no users left of the worker which asynchronously removes
a repository's on-disk remotes given that we only use in-memory remotes
nowadays. While we can now remove the infrastructure to asynchronously
remove remotes, we still need to retain the worker to make sure all jobs
are drained on an upgrade. The worker's logic is stubbed out though to
just do nothing: the worst that can happen with this is that we retain
on-disk remotes in Gitaly storages, but Gitaly will eventually clean
them up via repository maintenance anyway.