Run full rspec jobs on automated MRs
To ensure the full rspec job is run on automated MRs,
there is a new rule `if-automated-merge-reqeust`
added in the rspec job rules prior to
the rule `if-merge-request-not-approved`.

The order of these two rules are important to ensure the job
is created in the pipeline, instead of `when: never` taking effect.

To ensure the minimal job is not run on automated MRs,
a reverse rule is added to the minimal job rules
using `if-automated-merge-request` with `when: never`.