Document the actual behavior of remote push mirroring
The documentation of remote push mirroring doesn't reflect the actual
behavior of the feature. The documentation describes that by default the
mirroring would fail if there are any diverged refs in the mirror. This
doesn't match reality and the mirroring functionality by default force
pushes over the refs in the mirror. The keep divergent refs option then
actually avoids force pushing over the diverged refs but has no effect
on non-diverged refs.