Move ignore_column(s) from EE model to CE model
Prior to this change, models that exist in the CE and EE context and
used ignore_column(s) were incorrectly added to the EE model.
ignore_column(s) needs to be used in the CE model if the model exists in
CE and EE. If it only exists in EE, then the usage there is correct.

This will move the usages of ignore_column(s) to the CE model if a EE
model exists as well.