Add labels as search token in Jira issues list
It is added both as filtered search value and token. This
allows users to see labels as the selected token and they
can even select more labels in the filter bar if needed.

We are currently not using `fetchLabels` since the backend
for this is not ready yet.

Changelog: changed
EE: true