Disable creating user mentions during import
Creating a note causes a subtransaction every time `CacheMarkdownField`
attempts to refresh the issuable's user mentions. We disable it for now
to avoid large number of SQL queries. We are not using this table at the
moment for reads. https://gitlab.com/gitlab-org/gitlab/-/issues/21801
tracks the work necessary to make this work.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/336788

Changelog: performance