repository: Drop `#delete_config` function
The `#delete_config` function is not used anymore, and Gitaly has
deprecated direct access to the gitconfig in favor of more specialized
RPCs. Remove `#delete_config`.

Note that we do not yet delete `#set_config`: it's still in use to set
"gitlab.fullpath". This usecase will be addressed via a new
`SetFullPath()` RPC, at which point in time we can also remove
`#set_config`.