geo: Pass authentication header via `#fetch_remote`
While the `FetchRemote()` RPC allows the caller to specify an
authentication header, we still set it up manually in Geo via a call to
`#with_config`. Refactor the code to instead use the designated field
for this such that we don't needlessly call both `#set_config` and
`#delete_config`.

EE: true
Changelog: performance