Populate CommitCollection from target project/branch
Currently the commitsWithoutMergeCommits node list of the graphql
mergeRequest field is not populated if the _source_ project is not
accessible by the user. This is wrong and a break from the behaviour of
the existing REST API.

Fix this in the MergeRequestDiff object's load_commits() function by
populating the CommitCollection from the _target_ project and branch
instead. Otherwise the collection is determined to be inaccessible to
the user.

Fixes: https://gitlab.com/gitlab-org/gitlab/-/issues/335687
Changelog: fixed

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>