Rename functions to set repository's full path
In order to make it easy for an admin to see which project a given
repository belongs to with hashed storages, we always write the
project path (e.g. "gitlab-com/gitlab") into the repo's gitconfig. This
is done via `#write_repository_config()` and `#write_config()`, both of
which only support writing this specific config entry.

Rename both functions to make it clear that they only support writing
the path. Direct access to the gitconfig is going away soon anyway, so
these functions can never be extended to write additional config
entries.