Use type to detect password fields in integrations instead of name
When updating integration settings, password fields only need to be
filled in if the user wants to change them, otherwise the previous
values would be preserved.

This relied on clearing blank params with the name `password`, but some
integrations use a different name, such as `api_key` in the Datadog
integration. So instead we check for fields with `type: 'password'`,
which is also what the frontend does to decide how to render these
fields in the `DynamicField` component.

Changelog: fixed