Generate a Kubeconfig CI variable for configured cluster agents
If a project has associated cluster agents, a KUBECONFIG CI variable
will be provided to all CI jobs for the project. Each agent receieves
a corresponding Kubernetes context, which allows a job to connect to
the agent's cluster (via KAS) without further configuration.

The variable is only added if the job is not creating a deployment
to a cluster, because this could overwrite an existing variable.

https://gitlab.com/gitlab-org/gitlab/-/merge_requests/67089