Handle cache busting scenario when note author detail changes
Before when a note author which is part of a discussion has its
detail updated (e.g. name change, username change), the discussion
cache wont be busted.

To ensure that we show updated information when that happens, the
author's cache key is now included in
`Note#post_processed_cache_key` which we call in
`Discussion#cache_key`.

This way, when the author is updated, the discussion cache will be
invalidated as well.