Add rubocop rule for ignore_column(s) usages
Prior to this change, ignore rules for a columns were defined in CE or
EE models. But if a model exists in CE and EE, the ignore rule has to
be in the CE model. If the model only exists in EE, then it has to be
added there. A new rubocop rule will check for the correct location of
the ignore rule.