Move read_only/read_write back to Gitlab::Database
These methods were scoped to a connection in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/65262. A
discussion after the merge request was merged revealed these methods
aren't tied to a specific connection, and as such better belong in
Gitlab::Database.

I briefly considered moving them to the Gitlab module instead, as the
methods don't actually run any database queries. Unfortunately, this
isn't possible as this module is loaded too early to prepend it with an
EE module.