Finalize conversion to bigint for push_event_payloads
This is same change as
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/65795,
which had to be reverted because we hit deadlock when deploying to
GitLab.com.

This time we changed the lock order to be the same as in
EventCreateService#create_push_event, this should avoid the deadlock.

1. Ensure the migration of `event_id` to `event_id_convert_to_bigint` is
completed.
2. Check column types to make sure swapping is not already done
3. Copy indexes and FKs
4. Swap columns

See https://gitlab.com/gitlab-org/gitlab/-/issues/288005.

Changelog: other