Rename scan policies tab to policies
The Scan Policies page under the Security & Compliance tab is being
converted to Policies page where we will display a list of policies
and add the ability to select the project to scan the policies from.
This commit simply renames the menu and the breadcrumb to reflect
the changes from the migration.