Make Gitlab::BackgroundMigration.remaining work across queues
With worker routing, a Sidekiq queue can have jobs from multiple
workers. This meant `Gitlab::BackgroundMigration.remaining` needed to
account for that possibility. We do that in the same way as we already
did with the scheduled, retry, and dead sets: iterate through and look
for matching records.

We also fix other methods in this class (like `.steal`) while we're
doing this.